#21 Backend: Update api-gateway routes. Frontend: Add authorization by role, Add UI for home page, Add song list update after file loading

+++ frontend/src/app/components/songs/songs.component.ts
@@ -1,24 +1,24 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthenticationService } from "../../services/authentication.service";
-import { Router } from "@angular/router";
 import { SongService } from "../../services/song.service";
 import { Song } from "../../models/song";
 import { HttpResponse } from "@angular/common/http";
+import { retry } from "rxjs";
 
 const FILENAME_REGEX = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/;
 
 @Component({
     selector: 'app-songs',
     templateUrl: './songs.component.html',
-    styleUrls: ['./songs.component.css'],
     providers: [SongService]
 })
 export class SongsComponent implements OnInit {
     selectedFile: File;
     songList: Song[];
+    isSongLoading: boolean = false;
+    protected readonly Array = Array;
 
-    constructor(private authenticationService: AuthenticationService,
-        private router: Router,
+    constructor(public authService: AuthenticationService,
         private songService: SongService) {
     }
 
@@ -26,18 +26,25 @@ export class SongsComponent implements OnInit {
         this.updateSongList();
     }
 
-    onFileSelected(event: Event) {
-        this.selectedFile = event.target['files'][0];
-    }
-
     uploadFile() {
-        if (this.selectedFile) {
-            this.songService.uploadFile(this.selectedFile)
-                .subscribe(response => {
-                    console.log(response);
-                    this.updateSongList();
-                })
+        if (!this.authService.hasAuthority('ROLE_ADMIN')) {
+            return;
         }
+
+        if (!this.selectedFile) {
+            return;
+        }
+
+        this.isSongLoading = true;
+        this.songService.uploadFile(this.selectedFile)
+            .subscribe(response =>
+                this.songService.get(response['id'])
+                    .pipe(retry({delay: 500, count: 10}))
+                    .subscribe(song => {
+                        this.songList.push(song);
+                        this.isSongLoading = false;
+                    })
+            );
     }
 
     downloadFile(id: string) {
@@ -47,20 +54,30 @@ export class SongsComponent implements OnInit {
     }
 
     deleteSong(id: string) {
+        if (!this.authService.hasAuthority('ROLE_ADMIN')) {
+            return;
+        }
+
         this.songService.delete(id)
-            .subscribe(() => this.updateSongList())
+            .subscribe(() => {
+                this.songList = this.songList.filter(song => song.id !== id)
+            })
+    }
+
+    onFileSelected(event: Event) {
+        this.selectedFile = event.target['files'][0];
     }
 
     private updateSongList() {
-        this.songService.getSongList()
+        this.songService.getAll()
             .subscribe(songs => this.songList = songs);
     }
 
     private downloadBlob(response: HttpResponse<Blob>) {
         const objectUrl = URL.createObjectURL(response.body);
         const a = document.createElement('a');
-        a.href = objectUrl;
         a.download = this.parseFilename(response);
+        a.href = objectUrl;
         a.click();
 
         URL.revokeObjectURL(objectUrl);
@@ -71,8 +88,6 @@ export class SongsComponent implements OnInit {
         const matches = FILENAME_REGEX.exec(contentDisposition);
         return matches != null && matches[1]
             ? matches[1].replace(/['"]/g, '')
-            : 'audio';
+            : 'audio.mp3';
     }
-
-    protected readonly Array = Array;
 }
